@@ -36,39 +36,39 @@
     "load-grunt-tasks": "^3.5.2",
     "time-grunt": "^1.4.0",
     "tslint": "^3.15.1",
-    "typescript": "^2.0.3"
+    "typescript": "^2.0.10"
   },
   "dependencies": {
-    "@angular/common": "2.0.1",
-    "@angular/compiler": "2.0.1",
-    "@angular/core": "2.0.1",
-    "@angular/forms": "2.0.1",
-    "@angular/http": "2.0.1",
-    "@angular/platform-browser": "2.0.1",
-    "@angular/platform-browser-dynamic": "2.0.1",
-    "@angular/router": "3.0.1",
-    "@angular/upgrade": "2.0.1",
-    "angular2-notifications": "^0.4.0",
-    "bootstrap": "4.0.0-alpha.2",
+    "@angular/common": "2.2.3",
+    "@angular/compiler": "2.2.3",
+    "@angular/core": "2.2.3",
+    "@angular/forms": "2.2.3",
+    "@angular/http": "2.2.3",
+    "@angular/platform-browser": "2.2.3",
+    "@angular/platform-browser-dynamic": "2.2.3",
+    "@angular/router": "3.2.3",
+    "@angular/upgrade": "2.2.3",
+    "angular2-notifications": "0.4.0",
+    "bootstrap": "^3.3.7",
     "bourbon": "^4.2.7",
-    "chart.js": "^2.2.2",
+    "chart.js": "^2.4.0",
     "core-js": "^2.4.1",
     "datatables": "^1.10.12",
     "es6-shim": "^0.35.1",
-    "font-awesome": "^4.6.3",
+    "font-awesome": "^4.7.0",
     "grunt-sass": "^1.2.1",
     "grunt-ts": "^6.0.0-beta.3",
     "immutable": "^3.8.1",
-    "jquery": "^3.1.0",
+    "jquery": "^3.1.1",
     "jquery-ui-timepicker-addon": "^1.6.3",
     "jquery.inputmask": "^3.3.3",
-    "ng2-bootstrap": "1.1.5",
-    "primeng": "^1.0.0-beta.15",
+    "ng2-bootstrap": "^1.1.16",
+    "primeng": "^1.0.0",
     "primeui": "^4.1.15",
     "reflect-metadata": "^0.1.8",
     "rxjs": "5.0.0-beta.12",
-    "systemjs": "^0.19.38",
+    "systemjs": "^0.19.41",
     "underscore": "^1.8.3",
-    "zone.js": "^0.6.23"
+    "zone.js": "^0.7.2"
   }
 }
