@@ -6,9 +6,17 @@ using System.Threading.Tasks;
 
 namespace MIDAS.GBX.BusinessObjects
 {
-    public class ValidateOTP
+    public class ValidateOTP :GbObject
     {
         public OTP otp { get; set; }
         public User user { get; set; }
+
+        public override List<BusinessValidation> Validate<T>(T entity)
+        {
+            List<BusinessValidation> validations = new List<BusinessValidation>();
+            BusinessValidation validation = new BusinessValidation();
+
+            return validations;
+        }
     }
 }
