@@ -39,26 +39,6 @@ namespace MIDAS.GBX
             {
                 repo = new LocationRepository(context);
             }
-            if (typeof(T) == typeof(BO.MedicalProvider))
-            {
-                repo = new PasswordTokenRepository(context);
-            }
-            if (typeof(T) == typeof(BO.Specialty))
-            {
-                repo = new PasswordTokenRepository(context);
-            }
-            if (typeof(T) == typeof(BO.SpecialtyDetails))
-            {
-                repo = new PasswordTokenRepository(context);
-            }
-            if (typeof(T) == typeof(BO.Doctor))
-            {
-                repo = new PasswordTokenRepository(context);
-            }
-            if (typeof(T) == typeof(BO.DoctorSpeciality))
-            {
-                repo = new PasswordTokenRepository(context);
-            }
             if (typeof(T) == typeof(BO.Invitation))
             {
                 repo = new InvitationRepository(context);
@@ -66,6 +46,18 @@ namespace MIDAS.GBX
             if (typeof(T) == typeof(BO.SaveLocation))
             {
                 repo = new LocationRepository(context);
+            }
+            if (typeof(T) == typeof(BO.AddUser))
+            {
+                repo = new UserRepository(context);
+            }
+            if (typeof(T) == typeof(BO.ValidateOTP))
+            {
+                repo = new OTPRepository(context);
+            }
+            if (typeof(T) == typeof(BO.MedicalProvider))
+            {
+                repo = new MedicalProviderRepository(context);
             }
             return repo;
         }
