@@ -1,30 +1,21 @@
-import {Component, OnInit, ElementRef} from '@angular/core';
-import {Validators, FormGroup, FormBuilder} from '@angular/forms';
-import {Router} from '@angular/router';
-import {AppValidators} from '../../../utils/AppValidators';
-import {MedicalFacility} from '../../../models/medical-facility';
-import {MedicalFacilityDetail} from '../../../models/medical-facility-details';
-import {User} from '../../../models/user';
-import {Account} from '../../../models/account';
-import {Contact} from '../../../models/contact';
-import {Address} from '../../../models/address';
-import {SessionStore} from '../../../stores/session-store';
-import {NotificationsStore} from '../../../stores/notifications-store';
-import {Notification} from '../../../models/notification';
+import { Component, OnInit, ElementRef } from '@angular/core';
+import { Validators, FormGroup, FormBuilder } from '@angular/forms';
+import { Router } from '@angular/router';
+import { MedicalFacility } from '../../../models/medical-facility';
+import { SessionStore } from '../../../stores/session-store';
+import { NotificationsStore } from '../../../stores/notifications-store';
+import { Notification } from '../../../models/notification';
 import moment from 'moment';
-import {MedicalFacilityStore} from '../../../stores/medical-facilities-store';
-import {MedicalFacilityService} from '../../../services/medical-facility-service';
-import {StatesStore} from '../../../stores/states-store';
-import {StateService} from '../../../services/state-service';
+import { MedicalFacilityStore } from '../../../stores/medical-facilities-store';
+import { MedicalFacilityService } from '../../../services/medical-facility-service';
 
 @Component({
     selector: 'add-medical-facility',
     templateUrl: 'templates/pages/medical-facilities/add-medical-facility.html',
-    providers: [MedicalFacilityService, StateService, StatesStore, FormBuilder]
+    providers: [MedicalFacilityService, FormBuilder]
 })
 
 export class AddMedicalFacilityComponent implements OnInit {
-    states: any[];
     options = {
         timeOut: 3000,
         showProgressBar: true,
@@ -37,8 +28,6 @@ export class AddMedicalFacilityComponent implements OnInit {
     isSaveMedicalFacilityProgress = false;
 
     constructor(
-        private _stateService: StateService,
-        private _statesStore: StatesStore,
         private _medicalFacilitiesStore: MedicalFacilityStore,
         private fb: FormBuilder,
         private _router: Router,
@@ -47,67 +36,27 @@ export class AddMedicalFacilityComponent implements OnInit {
         private _elRef: ElementRef
     ) {
         this.medicalFacilityForm = this.fb.group({
-                name: ['', Validators.required],
-                prefix: ['', Validators.required],
-            contact: this.fb.group({
-                email: ['', [Validators.required, AppValidators.emailValidator]],
-                cellPhone: ['', [Validators.required]],
-                homePhone: [''],
-                workPhone: [''],
-                faxNo: ['']
-            }),
-            address: this.fb.group({
-                address1: [''],
-                address2: [''],
-                city: [''],
-                zipCode: [''],
-                state: [''],
-                country: ['']
-            })
+            name: ['', Validators.required],
+            npi: ['', Validators.required]
         });
 
         this.medicalFacilityFormControls = this.medicalFacilityForm.controls;
     }
 
     ngOnInit() {
-        this._stateService.getStates()
-            .subscribe(states => this.states = states);
     }
 
 
     saveMedicalFacility() {
         let medicalFacilityFormValues = this.medicalFacilityForm.value;
-        let medicalFacilityDetail = new MedicalFacilityDetail({
-            account: new Account({
-               id: this._sessionStore.session.account_id
-            }),
-            user: new User({
-                id: this._sessionStore.session.user.id
-            }),
-            medicalfacility: new MedicalFacility({
-                name: medicalFacilityFormValues.name,
-                prefix: medicalFacilityFormValues.prefix
-            }),
-            contactInfo: new Contact({
-                cellPhone: medicalFacilityFormValues.contact.cellPhone,
-                emailAddress: medicalFacilityFormValues.contact.email,
-                faxNo: medicalFacilityFormValues.contact.faxNo,
-                homePhone: medicalFacilityFormValues.contact.homePhone,
-                workPhone: medicalFacilityFormValues.contact.workPhone,
-            }),
-            address: new Address({
-                address1: medicalFacilityFormValues.address.address1,
-                address2: medicalFacilityFormValues.address.address2,
-                city: medicalFacilityFormValues.address.city,
-                country: medicalFacilityFormValues.address.country,
-                state: medicalFacilityFormValues.address.state,
-                zipCode: medicalFacilityFormValues.address.zipCode,
-            })
+        let medicalFacility = new MedicalFacility({
+            name: medicalFacilityFormValues.name,
+            npi: medicalFacilityFormValues.npi
         });
         this.isSaveMedicalFacilityProgress = true;
         let result;
 
-        result = this._medicalFacilitiesStore.addMedicalFacility(medicalFacilityDetail);
+        result = this._medicalFacilitiesStore.addMedicalFacility(medicalFacility);
         result.subscribe(
             (response) => {
                 let notification = new Notification({
