@@ -1,34 +1,22 @@
-import {Component, OnInit, ElementRef} from '@angular/core';
-import {Validators, FormGroup, FormBuilder} from '@angular/forms';
-import {Router, ActivatedRoute} from '@angular/router';
-import {AppValidators} from '../../../utils/AppValidators';
-import {MedicalFacility} from '../../../models/medical-facility';
-import {MedicalFacilityDetail} from '../../../models/medical-facility-details';
-import {User} from '../../../models/user';
-import {Account} from '../../../models/account';
-import {Contact} from '../../../models/contact';
-import {Address} from '../../../models/address';
-import {SessionStore} from '../../../stores/session-store';
-import {NotificationsStore} from '../../../stores/notifications-store';
-import {Notification} from '../../../models/notification';
+import { Component, OnInit, ElementRef } from '@angular/core';
+import { Validators, FormGroup, FormBuilder } from '@angular/forms';
+import { Router, ActivatedRoute } from '@angular/router';
+import { MedicalFacility } from '../../../models/medical-facility';
+import { SessionStore } from '../../../stores/session-store';
+import { NotificationsStore } from '../../../stores/notifications-store';
+import { Notification } from '../../../models/notification';
 import moment from 'moment';
-import {MedicalFacilityStore} from '../../../stores/medical-facilities-store';
-import {MedicalFacilityService} from '../../../services/medical-facility-service';
-import {StatesStore} from '../../../stores/states-store';
-import {StateService} from '../../../services/state-service';
+import { MedicalFacilityStore } from '../../../stores/medical-facilities-store';
+import { MedicalFacilityService } from '../../../services/medical-facility-service';
 
 @Component({
     selector: 'update-medical-facility',
     templateUrl: 'templates/pages/medical-facilities/update-medical-facility.html',
-    providers: [MedicalFacilityService, StateService, StatesStore, FormBuilder]
+    providers: [MedicalFacilityService, FormBuilder]
 })
 
 export class UpdateMedicalFacilityComponent implements OnInit {
-    medicalFacilityDetail = new MedicalFacilityDetail({});
     medicalfacility = new MedicalFacility({});
-    contactInfo = new Contact({});
-    address = new Address({});
-    states: any[];
     options = {
         timeOut: 3000,
         showProgressBar: true,
@@ -41,8 +29,6 @@ export class UpdateMedicalFacilityComponent implements OnInit {
     isSaveMedicalFacilityProgress = false;
 
     constructor(
-        private _stateService: StateService,
-        private _statesStore: StatesStore,
         private _medicalFacilitiesStore: MedicalFacilityStore,
         private fb: FormBuilder,
         private _router: Router,
@@ -55,13 +41,11 @@ export class UpdateMedicalFacilityComponent implements OnInit {
             let mfId: number = parseInt(routeParams.id);
             let result = this._medicalFacilitiesStore.fetchMedicalFacilityById(mfId);
             result.subscribe(
-                (medicalFacilityDetail: MedicalFacilityDetail) => {
-                    this.medicalfacility = medicalFacilityDetail.medicalfacility;
-                    this.contactInfo = medicalFacilityDetail.contactInfo;
-                    this.address = medicalFacilityDetail.address;
+                (medicalFacility: MedicalFacility) => {
+                    this.medicalfacility = medicalFacility;
                 },
                 (error) => {
-                    this._router.navigate(['/doctors']);
+                    this._router.navigate(['/medical-facilities']);
                 },
                 () => {
                 });
@@ -69,69 +53,30 @@ export class UpdateMedicalFacilityComponent implements OnInit {
 
         this.medicalFacilityForm = this.fb.group({
             name: ['', Validators.required],
-            prefix: ['', Validators.required],
-            contact: this.fb.group({
-                email: ['', [Validators.required, AppValidators.emailValidator]],
-                cellPhone: ['', [Validators.required]],
-                homePhone: [''],
-                workPhone: [''],
-                faxNo: ['']
-            }),
-            address: this.fb.group({
-                address1: [''],
-                address2: [''],
-                city: [''],
-                zipCode: [''],
-                state: [''],
-                country: ['']
-            })
+            npi: ['', Validators.required]
         });
 
         this.medicalFacilityFormControls = this.medicalFacilityForm.controls;
     }
 
     ngOnInit() {
-        this._stateService.getStates()
-            .subscribe(states => this.states = states);
     }
 
 
     updateMedicalFacility() {
         let medicalFacilityFormValues = this.medicalFacilityForm.value;
-        let medicalFacilityDetail = new MedicalFacilityDetail({
-            account: new Account({
-                id: this._sessionStore.session.account_id
-            }),
-            user: new User({
-                id: this._sessionStore.session.user.id
-            }),
-            medicalfacility: new MedicalFacility({
-                id: this.medicalFacilityDetail.medicalfacility.id,
-                name: medicalFacilityFormValues.name,
-                prefix: medicalFacilityFormValues.prefix
-            }),
-            contactinfo: new Contact({
-                cellPhone: medicalFacilityFormValues.contact.cellPhone,
-                emailAddress: medicalFacilityFormValues.contact.email,
-                faxNo: medicalFacilityFormValues.contact.faxNo,
-                homePhone: medicalFacilityFormValues.contact.homePhone,
-                workPhone: medicalFacilityFormValues.contact.workPhone,
-            }),
-            address: new Address({
-                address1: medicalFacilityFormValues.address.address1,
-                address2: medicalFacilityFormValues.address.address2,
-                city: medicalFacilityFormValues.address.city,
-                country: medicalFacilityFormValues.address.country,
-                state: medicalFacilityFormValues.address.state,
-                zipCode: medicalFacilityFormValues.address.zipCode,
-            })
+        let medicalFacility = new MedicalFacility({
+            name: medicalFacilityFormValues.name,
+            npi: medicalFacilityFormValues.npi,
+            id: this.medicalfacility.id
         });
         this.isSaveMedicalFacilityProgress = true;
         let result;
 
-        result = this._medicalFacilitiesStore.updateMedicalFacility(medicalFacilityDetail);
+        result = this._medicalFacilitiesStore.updateMedicalFacility(medicalFacility);
         result.subscribe(
             (response) => {
+                this.isSaveMedicalFacilityProgress = false;
                 let notification = new Notification({
                     'title': 'Medical facility updated successfully!',
                     'type': 'SUCCESS',
@@ -141,6 +86,7 @@ export class UpdateMedicalFacilityComponent implements OnInit {
                 this._router.navigate(['/medical-facilities']);
             },
             (error) => {
+                this.isSaveMedicalFacilityProgress = false;
                 let notification = new Notification({
                     'title': 'Unable to update Medical facility.',
                     'type': 'ERROR',
