@@ -1,5 +1,4 @@
 export enum SubscriptionPlan {
-    TRIAL = 0,
-    SINGLE_USER = 1,
-    VOLUME_LICENSE = 2
+    TRIAL = 1,
+    PRO = 2
 }
